+++ src/asyncmd/config.py
@@ -13,7 +13,7 @@
 # You should have received a copy of the GNU General Public License
 # along with asyncmd. If not, see <https://www.gnu.org/licenses/>.
 """
-This module contains the implementation of functions configuring asyncmd behavior.
+This module contains the implementation of functions configuring asyncmd resource usage.
 
 It also import the configuration functions for submodules (like slurm) to make
 them accessible to users in one central place.
@@ -22,12 +22,24 @@ import os
 import asyncio
 import logging
 import resource
+import typing
 
 
-from ._config import _GLOBALS, _SEMAPHORES, _OPT_SEMAPHORES
-from .trajectory.trajectory import _update_cache_type_for_all_trajectories
+from ._config import (_GLOBALS, _SEMAPHORES, _OPT_SEMAPHORES,
+                      _GLOBALS_KEYS, _SEMAPHORES_KEYS, _OPT_SEMAPHORES_KEYS,
+                      )
+from .trajectory.trajectory import (_update_cache_type_for_all_trajectories,
+                                    _deregister_h5py_cache_for_all_trajectories,
+                                    )
+from .trajectory.trajectory_cache import (
+        TrajectoryFunctionValueCacheInH5PY as _TrajectoryFunctionValueCacheInH5PY,
+        )
 # pylint: disable-next=unused-import
 from .slurm import set_slurm_settings, set_all_slurm_settings
+
+
+if typing.TYPE_CHECKING:  # pragma: no cover
+    import h5py
 
 
 logger = logging.getLogger(__name__)
@@ -62,7 +74,7 @@ def set_max_process(num: int | None = None, max_num: int | None = None) -> None:
             num = 2
     if max_num is not None:
         num = min((num, max_num))
-    _SEMAPHORES["MAX_PROCESS"] = asyncio.BoundedSemaphore(num)
+    _SEMAPHORES[_SEMAPHORES_KEYS.MAX_PROCESS] = asyncio.BoundedSemaphore(num)
 
 
 set_max_process()
@@ -121,7 +133,7 @@ def set_max_files_open(num: int | None = None, margin: int = 30) -> None:
     #       threads. The problem is that this semaphore will never be freed
     #       without any process getting a semaphore...
     semaval = int((num - margin) / 3)
-    _SEMAPHORES["MAX_FILES_OPEN"] = asyncio.BoundedSemaphore(semaval)
+    _SEMAPHORES[_SEMAPHORES_KEYS.MAX_FILES_OPEN] = asyncio.BoundedSemaphore(semaval)
 
 
 set_max_files_open()
@@ -147,9 +159,9 @@ def set_slurm_max_jobs(num: int | None) -> None:
     # pylint: disable-next=global-variable-not-assigned
     global _OPT_SEMAPHORES
     if num is None:
-        _OPT_SEMAPHORES["SLURM_MAX_JOB"] = None
+        _OPT_SEMAPHORES[_OPT_SEMAPHORES_KEYS.SLURM_MAX_JOB] = None
     else:
-        _OPT_SEMAPHORES["SLURM_MAX_JOB"] = asyncio.BoundedSemaphore(num)
+        _OPT_SEMAPHORES[_OPT_SEMAPHORES_KEYS.SLURM_MAX_JOB] = asyncio.BoundedSemaphore(num)
 
 
 set_slurm_max_jobs(num=None)
@@ -188,9 +200,9 @@ def set_trajectory_cache_type(cache_type: str,
     if (cache_type := cache_type.lower()) not in allowed_values:
         raise ValueError(f"Given cache type must be one of {allowed_values}."
                          + f" Was: {cache_type}.")
-    if _GLOBALS.get("TRAJECTORY_FUNCTION_CACHE_TYPE", "not_set") != cache_type:
+    if _GLOBALS.get(_GLOBALS_KEYS.TRAJECTORY_FUNCTION_CACHE_TYPE, "not_set") != cache_type:
         # only do something if the new cache type differs from what we have
-        _GLOBALS["TRAJECTORY_FUNCTION_CACHE_TYPE"] = cache_type
+        _GLOBALS[_GLOBALS_KEYS.TRAJECTORY_FUNCTION_CACHE_TYPE] = cache_type
         _update_cache_type_for_all_trajectories(copy_content=copy_content,
                                                 clear_old_cache=clear_old_cache,
                                                 )
@@ -199,15 +211,26 @@ def set_trajectory_cache_type(cache_type: str,
 set_trajectory_cache_type("npz")
 
 
-def register_h5py_cache(h5py_group) -> None:
+def register_h5py_cache(h5py_group: "h5py.Group | h5py.File", copy_h5py: bool = False,
+                        copy_content: bool = True, clear_old_cache: bool = False,
+                        ) -> None:
     """
     Register a h5py file or group for CV value caching.
 
-    Note that this also sets the default cache type to "h5py", i.e. it calls
-    :func:`set_trajectory_cache_type` with ``cache_type="h5py"``.
+    Optionally copy over all cached values from the previously set h5py_cache(s),
+    also for :class:`Trajectory` objects that are currently not instantiated,
+    see the ``copy_h5py`` argument. If it is True all previously set caches will
+    be deregistered after copying (since their values are now available in newly
+    set cache also).
 
-    Note that a ``h5py.File`` is just a slightly special ``h5py.Group``, so you
-    can pass either. :mod:`asyncmd` will use either the file or the group as
+    Note that in case the trajectory cache type is currently not "h5py", this
+    function sets the cache type to "h5py", i.e. it calls :func:`set_trajectory_cache_type`
+    with ``cache_type="h5py"``.
+    The arguments ``copy_content`` and ``clear_old_cache`` are directly passed
+    to :func:`set_trajectory_cache_type`.
+
+    Note that a :class:`h5py.File` is just a slightly special :class:`h5py.Group`,
+    so you can pass either. :mod:`asyncmd` will use either the file or the group as
     the root of its own stored values.
     E.g. you will have ``h5py_group["asyncmd/TrajectoryFunctionValueCache"]``
     always pointing to the cached trajectory values and if ``h5py_group`` is
@@ -219,11 +242,80 @@ def register_h5py_cache(h5py_group) -> None:
     ----------
     h5py_group : h5py.Group or h5py.File
         The file or group to use for caching.
+    copy_h5py : bool, optional, by default False
+        Whether to copy over all cached values from the previously set h5py cache
+        (even for :class:`Trajectory` objects that are currently not instantiated).
+    copy_content : bool, optional
+        Whether to copy the current cache content to the new cache,
+        by default True
+    clear_old_cache : bool, optional
+        Whether to clear the old/previously set cache, by default False.
     """
     # pylint: disable-next=global-variable-not-assigned
     global _GLOBALS
-    _GLOBALS["H5PY_CACHE"] = h5py_group
-    set_trajectory_cache_type(cache_type="h5py")
+    if _GLOBALS.get(_GLOBALS_KEYS.TRAJECTORY_FUNCTION_CACHE_TYPE, "not_set") != "h5py":
+        # nothing to copy as h5py was not the old cache type
+        if h5py_group.file.mode == "r":
+            _GLOBALS[_GLOBALS_KEYS.H5PY_CACHE_READ_ONLY_FALLBACKS] = (
+                [h5py_group] + _GLOBALS.get(_GLOBALS_KEYS.H5PY_CACHE_READ_ONLY_FALLBACKS, [])
+                )
+        else:
+            _GLOBALS[_GLOBALS_KEYS.H5PY_CACHE] = h5py_group
+        set_trajectory_cache_type(cache_type="h5py", copy_content=copy_content,
+                                  clear_old_cache=clear_old_cache)
+    else:
+        # cache type already is h5py
+        if not copy_h5py:
+            # but we dont want to copy, so just set the cache
+            if h5py_group.file.mode == "r":
+                _GLOBALS[_GLOBALS_KEYS.H5PY_CACHE_READ_ONLY_FALLBACKS] = (
+                    [h5py_group] + _GLOBALS.get(_GLOBALS_KEYS.H5PY_CACHE_READ_ONLY_FALLBACKS, [])
+                    )
+            else:
+                _GLOBALS[_GLOBALS_KEYS.H5PY_CACHE] = h5py_group
+        else:
+            if h5py_group.file.mode == "r":
+                raise ValueError("The passed h5py.File/h5py.Group is open in read-only "
+                                 "mode, but ``copy_h5py=True`` was passed. Can not "
+                                 f"copy because the file ({h5py_group}) is not writeable!"
+                                 )
+            # copy the old groups content to the new cache
+            caches_to_copy = _GLOBALS.get(_GLOBALS_KEYS.H5PY_CACHE_READ_ONLY_FALLBACKS, [])
+            if _GLOBALS.get(_GLOBALS_KEYS.H5PY_CACHE, None) is not None:
+                caches_to_copy += [_GLOBALS[_GLOBALS_KEYS.H5PY_CACHE]]
+            for cache in caches_to_copy:
+                _TrajectoryFunctionValueCacheInH5PY.add_values_for_all_trajectories(
+                    src_h5py_cache=cache, dst_h5py_cache=h5py_group,
+                    )
+            # and set the cache
+            _GLOBALS[_GLOBALS_KEYS.H5PY_CACHE] = h5py_group
+            # also empty the fallback cache list since all they contain should now
+            # be in the main cache
+            _GLOBALS[_GLOBALS_KEYS.H5PY_CACHE_READ_ONLY_FALLBACKS] = []
+        # although cache type was already h5py, update the cache type for all
+        # trajectories in existence to make sure they use the now set caches
+        _update_cache_type_for_all_trajectories(copy_content=copy_content,
+                                                clear_old_cache=clear_old_cache)
+
+
+def deregister_h5py_cache(h5py_group: "h5py.Group | h5py.File"):
+    """
+    Deregister a given h5py_group from use as a cache for trajectory function values.
+
+    Also deregisters the given h5py_group from all :class:`asyncmd.Trajectory` objects
+    currently in existence.
+
+    Parameters
+    ----------
+    h5py_group : h5py.Group | h5py.File
+        The h5py_group to deregister.
+    """
+    if _GLOBALS.get(_GLOBALS_KEYS.TRAJECTORY_FUNCTION_CACHE_TYPE, "not_set") == "h5py":
+        _deregister_h5py_cache_for_all_trajectories(h5py_group=h5py_group)
+    if _GLOBALS.get(_GLOBALS_KEYS.H5PY_CACHE, None) is h5py_group:
+        _GLOBALS[_GLOBALS_KEYS.H5PY_CACHE] = None
+    if h5py_group in _GLOBALS.get(_GLOBALS_KEYS.H5PY_CACHE_READ_ONLY_FALLBACKS, []):
+        _GLOBALS[_GLOBALS_KEYS.H5PY_CACHE_READ_ONLY_FALLBACKS].remove(h5py_group)
 
 
 def show_config() -> None: